requirements.txt
Django==1.11.7
Pillow==4.3.0
django-crispy-forms==1.7.0
